Boring type-mismatch fixups (s/unsigned int/NSUInteger/).
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@5972 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/NSObject+CoreObject.h
@@ -24,7 +24,7 @@
 - (ETUUID *) UUID;
 
 /** Returns a hash based on the UUID. */
-- (unsigned int) hash;
+- (NSUInteger) hash;
 
 /** Returns the current version of the instance. This version represents a 
     revision in the object history.
